Don't force mobile node user to pass glue code for all exposed service types, use separate specific methods instead
