+++ mobile/mysterium/openvpn_connection_setup.go
@@ -22,7 +22,27 @@ import (
 	"github.com/mysteriumnetwork/go-openvpn/openvpn3"
 	"github.com/mysteriumnetwork/node/client/stats"
 	"github.com/mysteriumnetwork/node/core/connection"
+	"github.com/mysteriumnetwork/node/services/openvpn"
+	"github.com/mysteriumnetwork/node/services/openvpn/session"
 )
+
+type sessionWrapper struct {
+	session *openvpn3.Session
+}
+
+func (wrapper *sessionWrapper) Start() error {
+
+	wrapper.session.Start()
+	return nil
+}
+
+func (wrapper *sessionWrapper) Stop() {
+	wrapper.session.Stop()
+}
+
+func (wrapper *sessionWrapper) Wait() error {
+	return wrapper.session.Wait()
+}
 
 type statsUpdater interface {
 	Save(stats stats.SessionStats)
@@ -63,5 +83,47 @@ func (adapter channelToCallbacksAdapter) OnStats(openvpnStats openvpn3.Statistic
 	adapter.statsUpdater.Save(stats.SessionStats{
 		BytesSent:     uint64(openvpnStats.BytesOut),
 		BytesReceived: uint64(openvpnStats.BytesIn),
+	})
+}
+
+// Openvpn3TunnelSetup is alias for openvpn3 tunnel setup interface exposed to Android/iOS interop
+type Openvpn3TunnelSetup openvpn3.TunnelSetup
+
+// OverrideOpenvpnConnection replaces default openvpn connection factory with mobile related one
+func (mobNode *MobileNode) OverrideOpenvpnConnection(tunnelSetup Openvpn3TunnelSetup) {
+	mobNode.di.ConnectionRegistry.Register("openvpn", func(options connection.ConnectOptions, channel connection.StateChannel) (connection.Connection, error) {
+
+		vpnClientConfig, err := openvpn.NewClientConfigFromSession(options.SessionConfig, "", "")
+		if err != nil {
+			return nil, err
+		}
+
+		profileContent, err := vpnClientConfig.ToConfigFileContent()
+		if err != nil {
+			return nil, err
+		}
+
+		config := openvpn3.NewConfig(profileContent)
+		config.GuiVersion = "govpn 0.1"
+		config.CompressionMode = "asym"
+		config.ConnTimeout = 0 //essentially means - reconnect forever (but can always stopped with disconnect)
+
+		signer := mobNode.di.SignerFactory(options.ConsumerID)
+
+		username, password, err := session.SignatureCredentialsProvider(options.SessionID, signer)()
+		if err != nil {
+			return nil, err
+		}
+
+		credentials := openvpn3.UserCredentials{
+			Username: username,
+			Password: password,
+		}
+
+		session := openvpn3.NewMobileSession(config, credentials, channelToCallbacks(channel, mobNode.di.StatsKeeper), tunnelSetup)
+
+		return &sessionWrapper{
+			session: session,
+		}, nil
 	})
 }
