--- mobile/mysterium/wrapper.go
@@ -1,38 +1,0 @@
-/*
- * Copyright (C) 2018 The "MysteriumNetwork/node" Authors.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-package mysterium
-
-import "github.com/mysteriumnetwork/go-openvpn/openvpn3"
-
-type sessionWrapper struct {
-	session *openvpn3.Session
-}
-
-func (wrapper *sessionWrapper) Start() error {
-
-	wrapper.session.Start()
-	return nil
-}
-
-func (wrapper *sessionWrapper) Stop() {
-	wrapper.session.Stop()
-}
-
-func (wrapper *sessionWrapper) Wait() error {
-	return wrapper.session.Wait()
-}
